Renamed MicrometerMetricsProvider to MicrometerTaggedMetricsProvider
